Conflicts:
	fortscale/fortscale-ml/src/main/java/fortscale/ml/model/cache/ModelsCacheService.java
	fortscale/fortscale-ml/src/main/java/fortscale/ml/scorer/algorithms/CategoryRarityModelScorerAlgorithm.java
	fortscale/fortscale-streaming/src/main/java/fortscale/streaming/service/model/ModelsCacheServiceSamza.java

+++ fortscale/fortscale-ml/src/main/java/fortscale/ml/scorer/algorithms/CategoryRarityModelScorerAlgorithm.java
@@ -47,7 +47,7 @@ public class CategoryRarityModelScorerAlgorithm {
         Assert.isTrue(maxNumOfRareFeatures >= 0, String.format("maxNumOfRareFeatures must be >= 0: %d", maxNumOfRareFeatures));
     }
 
-    public double calculateScore(int featureCount, CategoryRarityModel model) {
+    public double calculateScore(long featureCount, CategoryRarityModel model) {
         if(model==null) {
             return 0D;
         }
@@ -62,7 +62,7 @@ public class CategoryRarityModelScorerAlgorithm {
             numRareEvents += (i + 1) * buckets[i];
             numDistinctRareFeatures += buckets[i];
         }
-        for (int i = featureCount; i < featureCount + maxRareCount; i++) {
+        for (int i = (int) featureCount; i < featureCount + maxRareCount; i++) {
             double commonnessDiscount = calcCommonnessDiscounting(i - featureCount + 2, maxRareCount );
             numRareEvents += (i + 1) * buckets[i] * commonnessDiscount;
             numDistinctRareFeatures += buckets[i] * commonnessDiscount;
